Move debug related endpoints to seperate service part
